sua dto

+++ BUS/ViewModel/CreateOrderViewModel.cs
@@ -10,10 +10,10 @@ namespace BUS.ViewModel
     {
       
         public Guid UserID { get; set; }
-        public float AmountPay { get; set; }
-        public float PayingCustomer { get; set; }
+        public Int64 AmountPay { get; set; }
+        public Int64 PayingCustomer { get; set; }
         public string Description { get; set; }
-        public float Payments { get; set; }
+        public Int64 Payments { get; set; }
         
         public List<CartViewModel> CartViewModel { get; set; }
