+++ BUS/ViewModel/UpdateProductDetail.cs
@@ -11,8 +11,8 @@ namespace BoutiqueFashionFirstCode.ViewModel
         public Guid VariantId { get; set; }
         public string productsName { get; set; }
         public string skuId { get; set; }
-        public float importPrice { get; set; }
-        public float price { get; set; }
+        public Int64 importPrice { get; set; }
+        public Int64 price { get; set; }
         public int quantity { get; set; }
         public ICollection<ImageValueObject>? Images { get; set; }
         public List<OptionDto>? option { get; set; }
